Refactoring api files
