+++ frontend/app/api/index.ts
@@ -1,5 +1,4 @@
-export * from "./accountDetail";
-export * from "./accountSummary";
+export * from "./accounts";
 export * from "./banks";
 export * from "./client";
 export * from "./environment";
