+++ frontend/app/api/sync/index.ts
@@ -1,4 +1,4 @@
-import { prisma } from "./client";
+import { prisma } from "../client";
 
 export type SyncStatus =
   | "notsynced"
